% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_functions.R
\name{calc.LMG.contribution}
\alias{calc.LMG.contribution}
\title{A function that calculates the LMG-contribution given a subset model. The contribution is calculated as discribed by Byhring(2019);
Relative variable importance in linear regression models with random intercept term}
\usage{
calc.LMG.contribution(df, var1, var2, var3, var4)
}
\arguments{
\item{df}{A data.frame containing the fixed effect variables}

\item{var1}{The response variable}

\item{var2}{The random intercept variable}

\item{var3}{The variable of interest, that is the one we are interested in calculating the importance of.}

\item{var4}{The number of fixed effects in the full model.}
}
\value{
A numeric value containing the LMG contribution of a subset model.
}
\description{
A function that calculates the LMG-contribution given a subset model. The contribution is calculated as discribed by Byhring(2019);
Relative variable importance in linear regression models with random intercept term
}
