% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_functions.R
\name{calc.R2}
\alias{calc.R2}
\title{A function that calculates the explained variance, R^2, of a random intercept model with one random effect.
The method is described an article by Nakagawa and Schielzeth(2013);
A general and simple method for obtaining R^2 from generalized linear mixed-effects models.}
\usage{
calc.R2(LMM, marginal.r2 = TRUE)
}
\arguments{
\item{LMM}{The lmer object you want to calculate the explained variance of}

\item{marginal.r2}{whether the marginal(only regard fixed effect variance as explained variance)
or the conditional(also regards the random effect variance as explained) R^2 should be computed, defaults to TRUE which computes the marginal}
}
\value{
A numeric value. The estimated explained variance of the input model.
}
\description{
A function that calculates the explained variance, R^2, of a random intercept model with one random effect.
The method is described an article by Nakagawa and Schielzeth(2013);
A general and simple method for obtaining R^2 from generalized linear mixed-effects models.
}
\keyword{explained}
\keyword{variance}
