% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_functions.R
\name{calc.relimp.lmm}
\alias{calc.relimp.lmm}
\title{A function that calculate the relative importance of all fixed effects and random effects in random intercept models created with the lme4 package.
It is currently limited to one random effect. The method is an extension of one proposed by Lindeman, Marenda and Gold(1980).
The method is described and reviewd by Groemping(2007);
A general and simple method for obtaining R^2 from generalized linear mixed-effects models}
\usage{
calc.relimp.lmm(lmm.obj, response.name)
}
\arguments{
\item{lmm.obj}{An lmer object with one random intercept term}

\item{response.name}{The name of the response variable as a string}
}
\value{
A numeric value containing the LMG contribution of a subset model.
}
\description{
A function that calculate the relative importance of all fixed effects and random effects in random intercept models created with the lme4 package.
It is currently limited to one random effect. The method is an extension of one proposed by Lindeman, Marenda and Gold(1980).
The method is described and reviewd by Groemping(2007);
A general and simple method for obtaining R^2 from generalized linear mixed-effects models
}
\keyword{importance}
\keyword{variable}
