% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_functions.R
\name{create.ordering}
\alias{create.ordering}
\title{A function creates a matrix of orderings for which parameters that are to be included in the calculation of the average increase in R^2
The rows of the matrix denotes which parameters that are to be included in the model}
\usage{
create.ordering(n)
}
\arguments{
\item{n}{number of parameters that is going to be permuted}
}
\value{
A matrix
}
\description{
A function creates a matrix of orderings for which parameters that are to be included in the calculation of the average increase in R^2
The rows of the matrix denotes which parameters that are to be included in the model
}
