% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_functions.R
\name{get.subset}
\alias{get.subset}
\title{A function creates a subset of the full df based on a logical list.}
\usage{
get.subset(ordering, df)
}
\arguments{
\item{ordering}{A binary list}

\item{df}{The data.frame to create a subset of}
}
\value{
A subset data.frame of df
}
\description{
A function creates a subset of the full df based on a logical list.
}
